// src/stores/types/ProductImageCreate.ts
export interface ProductImageCreate_productImageCreate_errors {
  __typename: "ProductError";
  code: ProductErrorCode;
  field: string | null;
}
// ...
export interface ProductImageCreate_productImageCreate_product {
  __typename: "Product";
  id: string;
  attributes: ProductImageCreate_productImageCreate_product_attributes[];
  productType: ProductImageCreate_productImageCreate_product_productType;
  name: string;
  description: any;
  seoTitle: string | null;
  seoDescription: string | null;
  category: ProductImageCreate_productImageCreate_product_category | null;
  collections: (ProductImageCreate_productImageCreate_product_collections | null)[] | null;
  basePrice: ProductImageCreate_productImageCreate_product_basePrice | null;
  margin: ProductImageCreate_productImageCreate_product_margin | null;
  purchaseCost: ProductImageCreate_productImageCreate_product_purchaseCost | null;
  isAvailable: boolean | null;
  isPublished: boolean;
  chargeTaxes: boolean;
  publicationDate: any | null;
  pricing: ProductImageCreate_productImageCreate_product_pricing | null;
  images: (ProductImageCreate_productImageCreate_product_images | null)[] | null;
  variants: (ProductImageCreate_productImageCreate_product_variants | null)[] | null;
}

export interface ProductImageCreate_productImageCreate {
  __typename: "StoreImageCreate";
  storeErrors: ProductImageCreate_productImageCreate_errors[];
  store: ProductImageCreate_productImageCreate_product | null;
}

export interface ProductImageCreate {
  storeImagecreate: ProductImageCreate_productImageCreate | null;
}

export interface ProductImageCreateVariables {
  input: {
    store: string;
    image: any;
    alt?: string | null;
  }
}